+++ src/main/java/com/mdah/blog/controller/FolderController.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import javax.servlet.http.HttpSession;
+
 @Controller
 @RequestMapping("folder")
 public class FolderController {
@@ -13,13 +15,19 @@ public class FolderController {
     FolderDao folderDao;
 
     @RequestMapping("new")
-    public String newFolder(Folder folder) {
+    public String newFolder(Folder folder, HttpSession session) {
+        if (null == session.getAttribute("HasPermission")) {
+            return "pass";
+        }
         folderDao.save(folder);
         return "redirect:/index";
     }
 
     @RequestMapping("delete")
-    public String deleteFolder(Integer id) {
+    public String deleteFolder(Integer id, HttpSession session) {
+        if (null == session.getAttribute("HasPermission")) {
+            return "pass";
+        }
         folderDao.deleteById(id);
         return "redirect:/config";
     }
